Added conditional compilation of shortlived pool
