+++ pool.go
@@ -3,8 +3,6 @@ package bytebufferpool
 import (
 	"sort"
 	"sync/atomic"
-
-	"github.com/gallir/shortlivedpool"
 )
 
 const (
@@ -17,24 +15,6 @@ const (
 	calibrateCallsThreshold = 42000
 	maxPercentile           = 0.95
 )
-
-// Pool represents byte buffer pool.
-//
-// Distinct pools may be used for distinct types of byte buffers.
-// Properly determined byte buffer types with their own pools may help reducing
-// memory waste.
-type Pool struct {
-	calls       [steps]uint64
-	calibrating uint64
-
-	defaultSize uint64
-	maxSize     uint64
-
-	minBitSize uint64
-	minSize    uint64
-
-	pool shortlivedpool.Pool
-}
 
 var defaultPool Pool
